Refactoring and applying changes
