+++ torchmetrics/functional/text/bert.py
@@ -40,27 +40,27 @@ def bertscore(
     useful for evaluating different language generation tasks.
 
     Args:
-        - :param: `predictions` (list of str): candidate sentences
-        - :references: `refs` (list of str): reference sentences
-        - :param: `model_type` (str): bert specification, default using the suggested
+        - `predictions` (list of str): candidate sentences
+        - `references: `refs` (list of str): reference sentences
+        - `model_type` (str): bert specification, default using the suggested
                   model for the target langauge; has to specify at least one of
                   `model_type` or `lang`
-        - :param: `num_layers` (int): the layer of representation to use.
+        - `num_layers` (int): the layer of representation to use.
                   default using the number of layer tuned on WMT16 correlation data
-        - :param: `verbose` (bool): turn on intermediate status update
-        - :param: `idf` (bool or dict): use idf weighting, can also be a precomputed idf_dict
-        - :param: `device` (str): on which the contextual embedding model will be allocated on.
+        - `verbose` (bool): turn on intermediate status update
+        - `idf` (bool or dict): use idf weighting, can also be a precomputed idf_dict
+        - `device` (str): on which the contextual embedding model will be allocated on.
                   If this argument is None, the model lives on cuda:0 if cuda is available.
-        - :param: `nthreads` (int): number of threads
-        - :param: `batch_size` (int): bert score processing batch size
-        - :param: `lang` (str): language of the sentences; has to specify
+        - `nthreads` (int): number of threads
+        - `batch_size` (int): bert score processing batch size
+        - `lang` (str): language of the sentences; has to specify
                   at least one of `model_type` or `lang`. `lang` needs to be
                   specified when `rescale_with_baseline` is True.
-        - :param: `rescale_with_baseline` (bool): rescale bertscore with pre-computed baseline
-        - :param: `baseline_path` (str): customized baseline file
+        - `rescale_with_baseline` (bool): rescale bertscore with pre-computed baseline
+        - `baseline_path` (str): customized baseline file
 
     Returns:
-        - :param: `(P, R, F)`: each is of shape (N); N = number of input
+        - `(P, R, F)`: each is of shape (N); N = number of input
                       candidate reference pairs. if returning hashcode, the
                       output will be ((P, R, F), hashcode). If a candidate have
                       multiple references, the returned score of this candidate is
